Don't run task `Set up db user password` when in check mode
Setting passwords in Debconf always creates a change, which is annoying
when running the role in check mode in order to track real changes.
That's why it's better to don't run this particular task at all in check
mode.

This change makes the role idempotent at least in check mode.